+++ src/Facades/Vspay.php
@@ -10,6 +10,7 @@ use Topvendor\Vspay\Resources\Checkout;
 use Topvendor\Vspay\Resources\Gateway;
 use Topvendor\Vspay\Resources\Payments;
 use Topvendor\Vspay\Resources\Refunds;
+use Topvendor\Vspay\Resources\Uz;
 use Topvendor\Vspay\Webhooks\WebhookVerifier;
 
 /**
@@ -17,9 +18,12 @@ use Topvendor\Vspay\Webhooks\WebhookVerifier;
  * @method static Refunds refunds()
  * @method static Gateway gateway()
  * @method static Checkout checkout()
+ * @method static Uz uz()
  * @method static WebhookVerifier webhooks()
  * @method static Config config()
  * @method static Response post(string $path, array $body)
+ * @method static Response postProvider(string $path, array $body)
+ * @method static Response getProvider(string $path)
  *
  * @see VspayClient
  */